chore: modify example router setting for CI
