+++ examples/basic-auth/src/pages/Home/index.tsx
@@ -78,8 +78,4 @@ const HomePage = () => {
   );
 };
 
-HomePage.pageConfig = {
-  auth: ['admin']
-};
-
 export default HomePage;
