+++ examples/basic-nested-store/src/pages/About/index.tsx
@@ -35,8 +35,4 @@ const About = (props) => {
   );
 };
 
-About.pageConfig = {
-  title: 'About'
-};
-
 export default About;
