+++ examples/basic-ssr-with-lazy-load/src/pages/About/index.tsx
@@ -17,8 +17,4 @@ const About = (props) => {
   );
 };
 
-About.getInitialProps = async () => {
-  return { title: 'About Page...' };
-};
-
 export default About;
