+++ examples/basic-ssr-with-lazy-load/src/routes.ts
@@ -14,11 +14,31 @@ export default [
         path: '/home',
         exact: true,
         component: Home,
+        getInitialProps: async () => {
+          // const res = await request('/profile');
+          const res = {
+            data: {
+              profile: {
+                id: 10001,
+                name: 'Jack Ma',
+                edu: 'Hangzhou Normal University',
+                address: 'Hangzhou'
+              },
+              title: 'Home Page...',
+              content: 'Home Content...',
+              description: 'Home Description...'
+            }
+          };
+          return { ...res.data, title: 'Home Page...' };
+        }
       },
       {
         path: '/about',
         exact: true,
-        component: About
+        component: About,
+        getInitialProps: async () => {
+          return { title: 'About Page...' };
+        }
       },
       {
         path: '/',
