+++ examples/basic-ssr/src/pages/About/index.tsx
@@ -18,8 +18,4 @@ const About = (props) => {
   );
 };
 
-About.getInitialProps = async () => {
-  return { title: 'About Page...' };
-};
-
 export default About;
