+++ examples/basic-ssr/src/pages/Dashboard/index.tsx
@@ -11,15 +11,4 @@ const Dashboard = (props) => {
   );
 };
 
-Dashboard.getInitialProps = async (ctx) => {
-  console.log('Dashboard ctx', ctx);
-
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve({ title: 'Dashboard Page xxxx...' });
-    }, 1 * 1000);
-  });
-  // return { title: 'Dashboard Page...' }
-};
-
 export default Dashboard;
