+++ examples/basic-ssr/src/pages/Home/index.tsx
@@ -37,22 +37,3 @@ export default function Home(props) {
     </main>
   );
 }
-
-Home.getInitialProps = async () => {
-  // const res = await request('/profile');
-  const res = {
-    data: {
-      profile: {
-        id: 10001,
-        name: 'Jack Ma',
-        edu: 'Hangzhou Normal University',
-        address: 'Hangzhou'
-      },
-      title: 'Home Page...',
-      content: 'Home Content...',
-      description: 'Home Description...'
-    }
-  };
-
-  return { ...res.data, title: 'Home Page...' };
-};
