+++ examples/basic-store/src/routes.ts
@@ -23,7 +23,10 @@ export default [
       },
       {
         path: '/about',
-        component: About
+        component: About,
+        pageConfig: {
+          title: 'About'
+        }
       },
       {
         path: '/detail',
